Update conversion.ml
added function flatten_or which writes expressions with only 'or's in an easier to read way, and a function or_distribution, which takes two fnc formulas p and q (their type is prop but they are written in normal conjunctive form) and commpute 'p or q'. These functions make it possible for line 76 to work.